added bin/s2graph-daemon.sh that launches daemon processes.
After doing 'activator clean package', the package for distribution will be prepared at ./target/deploy/, where you can execute commands like:

    bin/s2graph-daemon.sh start s2rest_play
    bin/s2graph-daemon.sh stop s2rest_play

To start or stop a daemon process for s2rest_play. The following will run the server in the foreground

    bin/s2graph-daemon.sh run s2rest_play

These files are internal scripts and will NOT be what end-users will be calling or modifying.

For that purpose, `start-s2graph.sh`, `stop-s2graph.sh`, etc. is to be added.