PostProcess.scala
The "select" field can affect the edge data fields to include in the query result.	
In my application , I used getEdges  operation, and I want get the total "props" result in the query result;
so I set "select":["props"] in the getEdges’s command  according to "Query Level Options",but it did not get the right result.
actually，the "select" field can affect the child filed of “poprs”，but can not  affect the total "props" field.